Package com.hivemq.extensions.discovery.s3 -> com.hivemq.extensions.cluster.discovery.s3
