+++ src/main/java/com/hivemq/extensions/cluster/discovery/s3/util/StringUtil.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2018-present HiveMQ GmbH
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.hivemq.extensions.cluster.discovery.s3.util;
+
+import com.hivemq.extension.sdk.api.annotations.Nullable;
+
+/**
+ * @author Abdullah Imal
+ */
+public final class StringUtil {
+
+    private StringUtil() {
+    }
+
+    public static boolean isNullOrBlank(final @Nullable String value) {
+        return value == null || value.isBlank();
+    }
+}
